+++ Sources/MapboxSearch/PublicAPI/RecordsProviderInteractor.swift
@@ -7,6 +7,10 @@ public protocol RecordsProviderInteractor {
     /// Engine would add it to the search index asap.
     func add(record: IndexableRecord)
     
+    /// Notify about records addition.
+    /// Engine would add it to the search index asap.
+    func add(records: [IndexableRecord])
+    
     /// Notify about record alteration.
     /// Engine would update search index for corresponding identifier asap.
     func update(record: IndexableRecord)
@@ -14,6 +18,10 @@ public protocol RecordsProviderInteractor {
     /// Notify about record deletion.
     /// Engine would drop search index for particular identifier.
     func delete(identifier: String)
+    
+    /// Notify about record deletion.
+    /// Engine would drop search index for  identifiers
+    func delete(identifiers: [String])
     
     /// Notify about database drop.
     /// Engine would drop layer related search indexes.