+++ Tests/MapboxSearchTests/RecordsProviderInteractorNativeCoreTests.swift
@@ -59,6 +59,27 @@ class RecordsProviderInteractorNativeCoreTests: XCTestCase {
         XCTAssertFalse(interactor.contains(identifier: record.id))
     }
     
+    // swiftlint:disable identifier_name
+    func testInteractorAddDeleteMultiRecords() {
+        let record_1 = IndexableRecordStub()
+        let record_2 = IndexableRecordStub()
+        let record_3 = IndexableRecordStub()
+        interactor.add(records: [record_1, record_2, record_3])
+        
+        XCTAssertEqual(layerStub.records.count, 3)
+        
+        XCTAssert(interactor.contains(identifier: record_1.id))
+        XCTAssert(interactor.contains(identifier: record_2.id))
+        XCTAssert(interactor.contains(identifier: record_3.id))
+        
+        interactor.delete(identifiers: [record_1.id, record_3.id])
+        
+        XCTAssertFalse(interactor.contains(identifier: record_1.id))
+        XCTAssertFalse(interactor.contains(identifier: record_3.id))
+        
+        XCTAssertTrue(interactor.contains(identifier: record_2.id))
+    }
+    
     func testInteractorDeleteAllRecords() {
         for _ in 0..<3 {
             interactor.add(record: IndexableRecordStub())